Refactor Store API service to be constructed simplier with just handler
